refactor: separate webpack 4 and 5 matching
