+++ packages/webcrack/src/unpack/webpack/unpack-webpack-5.ts
@@ -9,7 +9,6 @@ import {
   findAssignedEntryId,
   getModuleFunctions,
   modulesContainerMatcher,
-  webpackRequireFunctionMatcher,
 } from './common-matchers';
 import { WebpackModule } from './module';
 
@@ -29,7 +28,30 @@ export default {
   tags: ['unsafe'],
   scope: true,
   visitor(options = { bundle: undefined }) {
-    const { webpackRequire, containerId } = webpackRequireFunctionMatcher();
+    // Example: __webpack_modules__
+    const modulesId = m.capture(m.identifier());
+    const webpackRequire = m.capture(
+      m.functionDeclaration(
+        m.identifier(), // __webpack_require__
+        [m.identifier()], // moduleId
+        m.blockStatement(
+          m.anyList(
+            m.zeroOrMore(),
+            // Example: __webpack_modules__[moduleId](module, module.exports, __webpack_require__);
+            m.expressionStatement(
+              m.callExpression(
+                m.memberExpression(
+                  m.fromCapture(modulesId),
+                  m.identifier(),
+                  true,
+                ),
+              ),
+            ),
+            m.zeroOrMore(),
+          ),
+        ),
+      ),
+    );
     const container = modulesContainerMatcher();
 
     const matcher = m.blockStatement(
@@ -37,7 +59,7 @@ export default {
         m.zeroOrMore(),
         // Example: var __webpack_modules__ = { ... };
         m.variableDeclaration(undefined, [
-          m.variableDeclarator(containerId, container),
+          m.variableDeclarator(modulesId, container),
         ]),
         m.zeroOrMore(),
         webpackRequire,
